+++ darwin/exporter/formats/instance_mask.py
@@ -1,14 +1,13 @@
 import shutil
 import uuid
 from pathlib import Path
-from typing import Generator, List, Optional
+from typing import Generator
 
 import numpy as np
 from PIL import Image
-from upolygon import draw_polygon
 
 import darwin.datatypes as dt
-from darwin.utils import convert_polygons_to_sequences, get_progress_bar
+from darwin.utils import convert_polygons_to_mask, get_progress_bar, ispolygon
 
 
 def generate_instance_id(instance_ids, length=8):
@@ -35,20 +34,9 @@ def export(annotation_files: Generator[dt.AnnotationFile, None, None], output_di
                 cat = annotation.annotation_class.name
                 instance_id = generate_instance_id(instance_ids)
                 instance_ids.add(instance_id)
-                sequence = convert_polygons_to_sequences(annotation.data["path"], height, width)
-                mask = convert_polygons_to_mask(sequence, height, width) * 255
+                mask = convert_polygons_to_mask(annotation.data["path"], height, width) * 255
                 mask = Image.fromarray(mask.astype(np.uint8))
                 mask_id = f"{image_id}_{instance_id}"
                 mask.save(masks_dir / f"{mask_id}.png")
                 f.write(f"{image_id},{mask_id},{cat}\n")
         print(f"Dataset format saved at {output_dir}")
-
-
-def convert_polygons_to_mask(polygons, height, width):
-    mask = np.zeros((height, width)).astype(np.uint8)
-    draw_polygon(mask, polygons, 1)
-    return mask
-
-
-def ispolygon(annotation):
-    return annotation.annotation_type in ["polygon", "complex_polygon"]
